Metacode support and other major changes
git-svn-id: svn://81.25.47.234/lsfusion/plugin-idea@7 dcdcfa73-82ac-7b44-9532-e36433290fed

+++ src/com/simpleplugin/psi/references/impl/LSFMetaReferenceImpl.java
@@ -18,7 +18,7 @@ import org.jetbrains.annotations.NotNull;
 import java.util.ArrayList;
 import java.util.List;
 
-public abstract class LSFMetaReferenceImpl extends LSFFullNameReferenceImpl<LSFMetaDeclaration> implements LSFMetaReference {
+public abstract class LSFMetaReferenceImpl extends LSFFullNameReferenceImpl<LSFMetaDeclaration, LSFMetaDeclaration> implements LSFMetaReference {
 
     public LSFMetaReferenceImpl(@NotNull ASTNode node) {
         super(node);
@@ -57,7 +57,7 @@ public abstract class LSFMetaReferenceImpl extends LSFFullNameReferenceImpl<LSFM
 
     @Override
     public boolean isCorrect() {
-        return super.isCorrect() && getMetacodeUsage()!=null && getMetaCodeIdList()!=null;
+        return super.isCorrect() && getMetacodeUsage()!=null && getMetaCodeIdList()!=null && getNode().findChildByType(LSFTypes.SEMI)!=null;
     }
 
     @Override
@@ -91,11 +91,11 @@ public abstract class LSFMetaReferenceImpl extends LSFFullNameReferenceImpl<LSFM
     @Override
     public String getPreceedingTab() {
         ASTNode treePrev = getNode().getTreePrev();
-        if(treePrev==null) {
+/*        if(treePrev==null) {
             PsiElement parent = getParent();
             assert parent instanceof LSFStatements;
             treePrev = parent.getNode().getTreePrev(); // предполагается что тут statements будут
-        }
+        }*/
         if(treePrev!=null && LSFParserDefinition.isWhiteSpace(treePrev.getElementType())) // сохраним табуляцию
             return treePrev.getText();
         return "";
@@ -105,15 +105,22 @@ public abstract class LSFMetaReferenceImpl extends LSFFullNameReferenceImpl<LSFM
     public void setInlinedBody(LSFMetaCodeBody parsed) {
         LSFMetaCodeBody body = getMetaCodeBody();
         if (parsed == null || !isCorrect()) {
-            if (body != null)
+            if (body != null && body.isWritable())
                 body.delete();
         } else {
             if (body != null) {
-                body.replace(parsed);
+                if(!body.getText().equals(parsed.getText()))
+                    body.replace(parsed);
             } else {
-                getNode().addChild(parsed.getNode(), getMetaCodeIdList().getNode().getTreeNext().getTreeNext());
+                getNode().addChild(parsed.getNode(), getNode().findChildByType(LSFTypes.SEMI));
             }
         }
+    }
 
+    @Override
+    public void dropInlinedBody() {
+        LSFMetaCodeBody body = getMetaCodeBody();
+        if(body != null)
+            body.delete();
     }
 }
